suppress some warnings from clang trunk
git-svn-id: https://svn.r-project.org/R/trunk@77224 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -1130,7 +1130,7 @@ static size_t fifo_read(void *ptr, size_t size, size_t nitems,
     Rfifoconn this = con->private;
 
     /* uses 'size_t' for len */
-    if ((double) size * (double) nitems > SSIZE_MAX)
+    if ((double) size * (double) nitems > (double) SSIZE_MAX)
 	error(_("too large a block specified"));
     return read(this->fd, ptr, size * nitems)/size;
 }
@@ -1141,7 +1141,7 @@ static size_t fifo_write(const void *ptr, size_t size, size_t nitems,
     Rfifoconn this = con->private;
 
     /* uses 'size_t' for len */
-    if ((double) size * (double) nitems > SSIZE_MAX)
+    if ((double) size * (double) nitems > (double) SSIZE_MAX)
 	error(_("too large a block specified"));
     return write(this->fd, ptr, size * nitems)/size;
 }
@@ -2997,7 +2997,7 @@ static void text_init(Rconnection con, SEXP text, int type)
 	    (double) strlen(type == 1 ? translateChar(STRING_ELT(text, i))
 			    : ((type == 3) ?translateCharUTF8(STRING_ELT(text, i))
 			       : CHAR(STRING_ELT(text, i))) ) + 1;
-    if (dnc >= SIZE_MAX)
+    if (dnc >= (double) SIZE_MAX)
 	error(_("too many characters for text connection"));
     else nchars = (size_t) dnc;
     this->data = (char *) malloc(nchars+1);
